Version 1.1

+++ Interfaz/HistorialPedidos.cs
@@ -2,25 +2,26 @@
 using Helpers;
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
 using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace Code_test_L2
+namespace Interfaz
 {
     public partial class HistorialPedidos : Form
     {
         private Usuario user;
         private List<Pedido> listaPedidos;
+        /* 
+         * En los siguientes constructores la ventana se generará de una manera u otra dependiendo del usuario que
+         * haya ingresado.
+         */
         public HistorialPedidos(Cliente userCliente)
         {
             this.user = userCliente;
             InitializeComponent();
             btnMarcarDespachado.Visible = false;
+            this.BackColor = Color.SeaGreen;
+            dgvListadoPedidos.BackgroundColor = Color.BurlyWood;
         }
         public HistorialPedidos(Vendedor userVendedor)
         {
@@ -31,52 +32,71 @@ namespace Code_test_L2
 
         private void HistorialPedidos_Load(object sender, EventArgs e)
         {
-            cargar();
+            cargar(); // Al cargar la ventana, me actualiza el listado. 
         }
-
+        /* 
+         * El siguiente evento solo esta disponible para el Cliente y en el mismo podrá cancelar el pedido que realizó.
+         * Si el pedido ya fue despachado no podrá cancelarlo.
+         */
         private void btnCancelarPedido_Click(object sender, EventArgs e)
         {
-            Pedido seleccionado = (Pedido)dgvListadoPedidos.CurrentRow.DataBoundItem;
-            if (!(seleccionado.Estado == "Despachado"))
+            if (dgvListadoPedidos.CurrentRow != null)
             {
-                DialogResult resultado = MessageBox.Show("Esta seguro que quiere cancelar el pedido?", "Cancelando", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-                if (resultado == DialogResult.Yes)
+                Pedido seleccionado = (Pedido)dgvListadoPedidos.CurrentRow.DataBoundItem;
+                if (!(seleccionado.Estado == "Despachado"))
                 {
-                    if (user is Cliente cliente)
+                    DialogResult resultado = MessageBox.Show("Esta seguro que quiere cancelar el pedido?", "Cancelando", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                    if (resultado == DialogResult.Yes)
                     {
-                        Utilities.eliminarVentaVendedor(seleccionado.Vendedor);
-                        cliente.Pedidos.Remove(seleccionado);
+                        if (user is Cliente cliente)
+                        {
+                            Utilities.eliminarVentaVendedor(seleccionado.Vendedor);
+                            cliente.Pedidos.Remove(seleccionado);
+                        }
                     }
                 }
+                else
+                {
+                    MessageBox.Show("No puede cancelar un pedido que ya fue despachado. Espere que reciba el mismo a su domicilio para que pueda realizar la devolución.");
+                }
+                cargar();
             }
             else
             {
-                MessageBox.Show("No puede cancelar un pedido que ya fue despachado. Espere que reciba el mismo a su domicilio para que pueda realizar la devolución.");
+                MessageBox.Show("Debe seleccionar un pedido.");
             }
-            cargar();
-        }
-
+        }       
+         // El siguiente botón marca como despachado el pedido seleccionado. Solo disponible para usuario Vendedor.         
         private void btnMarcarDespachado_Click(object sender, EventArgs e)
         {
-            Pedido seleccionado = (Pedido)dgvListadoPedidos.CurrentRow.DataBoundItem;
-            DialogResult resultado = MessageBox.Show("Marcar como despachado?", "Despachando", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-            if (resultado == DialogResult.Yes)
+            if (dgvListadoPedidos.CurrentRow != null)
             {
-                seleccionado.Estado = "Despachado";
+                Pedido seleccionado = (Pedido)dgvListadoPedidos.CurrentRow.DataBoundItem;
+                DialogResult resultado = MessageBox.Show("Marcar como despachado?", "Despachando", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                if (resultado == DialogResult.Yes)
+                {
+                    seleccionado.Estado = "Despachado";
+                }
+                cargar();
             }
-            cargar();
+            else
+            {
+                MessageBox.Show("Debe seleccionar un pedido.");
+            }
         }
+        // Lo que hace cargar() es actualizar la lista de pedidos segun el usuario es vendedor o cliente.
         public void cargar()
         {
             if (user is Cliente cliente)
             {
                 dgvListadoPedidos.DataSource = null;
                 dgvListadoPedidos.DataSource = cliente.Pedidos;
-                dgvListadoPedidos.Columns["Estado"].Visible = false;
+                dgvListadoPedidos.Columns["Estado"].Visible = false; // Esta opcion solo lo puede ver el vendedor
                 dgvListadoPedidos.Columns["Cliente"].Visible = false;
             }
             else
             {
+                // Utilizo el método cargarPedidosSegunVendedor para que solo muestre los pedidos según el vendedor que ingreso.
                 listaPedidos = Utilities.cargarPedidosSegunVendedor(user.ApellidoNombre);
                 dgvListadoPedidos.DataSource = null;
                 dgvListadoPedidos.DataSource = listaPedidos;
